moved more stuff into utils- game works
